requirements.txt
scikit-image
scikit-video
torchvision
tqdm
msgpack>=1.0.0
termcolor
yacs
ipdb
opencv-python
opencv-contrib-python
numpy
torch
scipy
Pillow
imgaug
tensorboard
matplotlib
lmdb
detectron2
shapely
transformers
flask

fontTools
imagecorruptions